// frontend/src/api/generated/models/InsightsLabItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type InsightsLabItem = {
    test_name: string;
    value?: (string | null);
    unit?: (string | null);
    collected_at?: (string | null);
    is_abnormal?: boolean;
};